pyproject.toml
[tool.poetry]
name = "concrete-backup"
version = "0.1.0"
description = ""
authors = ["Wanja Wischmeier <wanja.wischmeier@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
PyQt6 = "^6.7.0"
psutil = "^6.0.0"
pyqt5 = "^5.15.11"
pyyaml = "^6.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
